Fix try-catch test (#134)
* Fix try-catch test

Fixes try-catch test, which was failing because gas was not being correcly decreased. We were decreasing the gas from the parent context frame without checking if there was a near call frame first.

* Remove load_program function

remove the load_program function which isn't used anymore

* Fix return unused gas on far calls

* Fix try-catch for some modes

Fixes try catch for 90% of all modes. They were failing because we were not properly reverting when deacreasing the opcode gas at the beginning of the execution

* Fix try-catch for all remaining modes

They were failing because we were not setting the register 1 as a pointer when returning from panic

* Fix merge error

* Refactor: move inexplicit panics to their own function

* Remove comments on handle_error

* Refactor: Add comments and rename  function
